fix: change ForwardBodyConfig.MaxSize from uint16 to int32
uint16 is not supported by go-to-protobuf; int32 aligns with other
Gateway API integer fields and fixes protobuf conversion for
ForwardBodyConfig.

Fixes #5068

+++ apis/v1/httproute_types.go
@@ -1791,7 +1791,7 @@ type ForwardBodyConfig struct {
 	//
 	// If 0, the body will not be sent to the authorization server.
 	// +optional
-	MaxSize uint16 `json:"maxSize,omitempty"`
+	MaxSize int32 `json:"maxSize,omitempty"`
 }
 
 // HTTPBackendRef defines how a HTTPRoute forwards a HTTP request.
